WIP moved front matter data dir to src
