+++ apps/coscrad-frontend/src/components/About/About.tsx
@@ -1,4 +1,4 @@
-import FrontMatter from '../../app/configurable-front-matter/frontMatterData/FrontMatter';
+import FrontMatter from '../../configurable-front-matter/frontMatterData/FrontMatter';
 import './About.module.scss';
 
 export interface AboutProps {
