+++ apps/coscrad-frontend/src/components/Header/Header.css
@@ -1,0 +1,3 @@
+h1 a:hover {
+    color: black;
+}
