+++ apps/coscrad-frontend/src/configurable-front-matter/frontMatterData/FrontMatter.ts
