+++ apps/coscrad-frontend/src/configurable-front-matter/frontMatterData/sample-front-matter.json
