+++ apps/coscrad-frontend/src/configurable-front-matter/validation/validateFrontMatter.ts
